Big change: list to set of attributes in Schema.
The order of inputs to commutative operators no longer matters.
Commutative CSE Elimination is now enabled, increasing CSE Elimination power.
Attributes now take the form of a proper class, rather than a string with special encoding.

There are two major changes to Spoof2Compiler.
The first is to check, at every Unbind, if a transpose is necessary. A transpose is necessary if the dimensions we are unbinding to differ from the dimensions we bound from below the unbind. The dimensions bound below the unbind are carried through the reconstruction procedure; nary and agg nodes change these, in particular.  The bound dimension map reflects the way Hop matrix dimensions are assigned to SPlan name dimensions.
The second is to check, at every root, whether the root changed orientation. We check the original input dimension that the root data dimensions connect to. If this changes from before SPlan rewriting to after, then we need to insert a transpose.

There are still problems within bigger scripts like the Autoencoder.
There are likely optimizations I can make, removing schema refreshes that are no longer necessary and such.