New factorization strategy in NormalFormExploreEq. Introduce EBlock.
EBlocks are equivalent sum-product blocks.
They are formed while factoring blocks, when there is a fork point
(in this case, multiple valid aggregation orders).

New * factorization pruning: when an aggName is independent of other aggNames
(it is not connected to another aggName), then factor it immediately.

Next up is the decision to factor common terms out of +.