Partial normal form extender. Fix evil bugs. Add many tests.
Test case 53 no longer factors as well as it ought to.
Need to introduce a limited form of renaming (BindUnify)
that does not put binds in the middle of blocks.

Don't factor out CSEs when the schemas of the identical inputs are aggregated differently.
Don't delete aggregates while factoring out CSEs from + by accident.