Create constant matrix multiplier in case of adding matrix to vector.
This patch solves a bug that occurs when a matrix is added to a vector
below a multiplication. The addition of the matrix and vector loses its
dimension during normal form rewriting.
We retain the dimension by multiplying the vector with a ones vector
to explicitly bring it up to the dimension of the matrix it is added to.
This constant vector is created via a DataGenOp.

Better ParseExplain.